# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de la Prueba GDA",
        "description": "Documentación API de la Prueba GDA",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/communes": {
            "get": {
                "tags": [
                    "communes"
                ],
                "summary": "Return communes",
                "description": "Return list of communes",
                "operationId": "e099cd2a86abfbf1c2c4bb55082274c1",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unsuccessful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "communes"
                ],
                "summary": "Create commune",
                "description": "Create commune",
                "operationId": "5dc073b0575d1d2611d1279907eadefb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    },
                                    "trash": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unsuccessful operation"
                    }
                }
            }
        },
        "/api/customers": {
            "post": {
                "tags": [
                    "customers"
                ],
                "summary": "Create customer",
                "description": "Create customer",
                "operationId": "7f198ffa8f163c1df17a6a586c665c49",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "dni": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "id_reg": {
                                        "type": "string"
                                    },
                                    "id_com": {
                                        "type": "string"
                                    },
                                    "trash": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unsuccessful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/regions": {
            "get": {
                "tags": [
                    "regions"
                ],
                "summary": "Returns regions",
                "description": "Returns list of regions",
                "operationId": "dfa5a1c014edc132075b845037714d19",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unsuccessful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "regions"
                ],
                "summary": "Create region",
                "description": "Create region",
                "operationId": "942a15b4b4b66748f4c2deffdfde9600",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    },
                                    "trash": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unsuccessful operation"
                    }
                }
            }
        },
...
        "/api/communes/{id}": {
            "delete": {
                "tags": [
                    "communes"
                ],
                "summary": "Delete communes",
                "description": "Delete communes",
                "operationId": "2a174a452c5790a0b8e3d7220cd9683a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unsuccessful operation"
                    },
                    "404": {
                        "description": "Registro no existe"
                    }
                }
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "customers"
                ],
                "summary": "Return customer",
                "description": "Return details of customer",
                "operationId": "1d8ff2232a99139f4f914c2b9c3c0135",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unsuccessful operation"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "customers"
                ],
                "summary": "Delete customer",
                "description": "Delete customer",
                "operationId": "05ea90a5af66f45a7e0576f863b709ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unsuccessful operation"
                    },
                    "404": {
                        "description": "Registro no existe"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/regions/{id}": {
            "delete": {
                "tags": [
                    "regions"
                ],
                "summary": "Delete region",
                "description": "Delete region",
                "operationId": "da0e2c6b775c3065cd1be568571b84a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unsuccessful operation"
                    },
                    "404": {
                        "description": "Registro no existe"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}